完成订单预约功能前后端交互

+++ PincheSystem-back/src/main/java/com/devhub/pinchesystemback/mapper/OrderMapper.java
@@ -11,7 +11,7 @@ import java.util.List;
 @Mapper
 public interface OrderMapper {
 
-    Long insert(Order order);
+    int insert(Order order);
 
     int deleteByOrderId(Long id);
